Made ping work on Windows without opening a command window
