+++ .gitignore
@@ -1,1 +1,2 @@
 ping.test.exe
+cmd/ping/ping*
